# application/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Programmer Notebook
  version: 1.0.0
servers:
  - url: https://0.0.0.0:8080/api/v2
...
paths:
...
  /teams:
    get:
      tags:
        - teams
      summary: Get all teams in database
      description: Получение всех команд
      operationId: getAllTeams

      security:
        - JWTAuth: []

      responses:
        "200":
          description: Success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamInfo'
        "403":
          description: Недостаточно прав
        "500":
          description: Internal error

    post:
      tags:
        - teams
      summary: Add new team
      operationId: addTeam

      security:
        - JWTAuth: []

      requestBody:
        description: Данные команды
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/TeamInfo'
        required: true

      responses:
        "200":
          description: Team successfully created
          content:
            'application/json':
              schema:
                type: object
                properties:
                  teamID:
                    type: integer
        "409":
          description: Team alredy exists
        "422":
          description: Invalid parameters
        "500":
          description: Internal error
  /teams/{id}:
    get:
      tags:
        - teams
      summary: Get team
      description: Получение одной команды
      operationId: getTeam

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID команды
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/TeamFullInfo'
        "401":
          description: Authorization error
        "403":
          description: Недостаточно прав
        "404":
          description: No such user
        "500":
          description: Internal error

    delete:
      tags:
        - teams
      summary: Delete team
      description: Удаление команды
      operationId: deleteTeam

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID команды
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Success
        "401":
          description: Authorization error
        "403":
          description: Недостаточно прав
        "404":
          description: No such command
        "500":
          description: Internal error

    patch:
      tags:
        - teams
      summary: Update team info
      description: Изменение данных команды
      operationId: updateTeam

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID команды
          required: true
          schema:
            type: integer

      requestBody:
        description: Новые данные команды
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/TeamInfo'
        required: true

      responses:
        "200":
          description: Success
        "401":
          description: Authorization error
        "403":
          description: Недостаточно прав
        "404":
          description: No such user
        "500":
          description: Internal error
  /teams/{id}/members:
    get:
      tags:
        - teams
      summary: Get team members
      description: "Получение членов команды или по ID, или по названию команды"
      operationId: getTeamMembers

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID команды
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserPublicInfo'
        "401":
          description: Authorization error
        "403":
          description: Недостаточно прав
        "404":
          description: No such user
        "500":
          description: Internal error

    post:
      tags:
        - teams
      summary: Add user to team
      description: Добавление пользователя в команду
      operationId: addUserToTeam

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID команды, в которую добавляем
          required: true
          schema:
            type: integer

      requestBody:
        description: ID пользователя, которого добавляем
        content:
          'application/json':
            schema:
              type: object
              properties:
                userID:
                  type: integer
        required: true

      responses:
        "200":
          description: Success
        "401":
          description: Authorization error
        "403":
          description: Недостаточно прав
        "404":
          description: No such user
        "500":
          description: Internal error
  /teams/{teamID}/members/{userID}:
    delete:
      tags:
        - teams
      summary: Delete user from team
      description: Удаление пользователя из команды
      operationId: deleteUserFromTeam

      security:
        - JWTAuth: []

      parameters:
        - name: teamID
          in: path
          description: ID команды, из которой удаляем
          required: true
          schema:
            type: integer

        - name: userID
          in: path
          description: ID пользователя, которого удаляем
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Success
        "401":
          description: Authorization error
        "403":
          description: Недостаточно прав
        "404":
          description: No such user
        "500":
          description: Internal error

      #####################################################################
      # Note API
      #####################################################################
...
components:
  schemas:
...
    UserPublicInfo:
      required:
        - id
        - fio
        - role
      type: object
      properties:
        id:
          type: integer
          format: int
          example: 10
        fio:
          type: string
          example: Иванов Иван Иванович
        role:
          type: integer
          description: "Роль пользователя (Админ, Читатель, Автор)"
          format: int
          example: 1
...
    TeamFullInfo:
      type: object
      required:
        - id
        - name
        - registrationDate
      properties:
        id:
          type: integer
          format: int
          example: 10
        name:
          type: string
          example: Фронтенд-команда
        registrationDate:
          type: string
          example: 01.01.2001
    TeamInfo:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int
          example: 10
        name:
          type: string
          example: Фронтенд-команда

    #####################################################################
    # Note schemas
    #####################################################################
...
  securitySchemes:
    JWTAuth:
      type: apiKey
      name: token
      in: header